Totcbecend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
sqlalchemy
psycopg2-binary
passlib[bcrypt]
pydantic[email]

fastapi
uvicorn
SQLAlchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic-settings
pydantic[email]
